Modify decoding to exhaustively produce output even in face of errors
This is needed to support error recovery modes as per WHATWG encoding spec.
DecodeError gets some more details: an error code telling whether the
error was caused by an invalid sequence or an unexpected end of input,
and the number of bytes to skip in the input when making error recovery.